feat: created a new helper for payment
