+++ src/helpers/convert-date.js
@@ -35,25 +35,6 @@ const formatDateMonthAndYear = (month, year) => {
 }
 
 /**
- * Check if the date is expired and return a string value 'Expired' if it is
- *
- * @param {number} month - The month
- * @param {number} year - The year
- * @returns {string} The string value
- */
-
-const getExpiredDate = (month, year) => {
-  const currentDate = new Date()
-  const currentYear = currentDate.getFullYear()
-  const currentMonth = currentDate.getMonth() + 1
-
-  const isExpiredByYear = year < currentYear
-  const isExpiredByMonth = year === currentYear && month < currentMonth
-
-  return isExpiredByYear || isExpiredByMonth ? 'Expired' : ''
-}
-
-/**
  * Formats a given date to ISO format without milliseconds.
  *
  * @param {Date} value - The date to be formatted.
@@ -102,4 +83,4 @@ const convertDateToLocalTimezone = (date, utcOffset) => {
   return formatToEndOfDayIsoDate(userRealDate)
 }
 
-export { convertValueToDate, convertDateToLocalTimezone, formatDateMonthAndYear, getExpiredDate }
+export { convertValueToDate, convertDateToLocalTimezone, formatDateMonthAndYear }
