+++ src/services/billing-services/list-payment-service.js
@@ -1,6 +1,7 @@
 import { AxiosHttpClientAdapter, parseHttpResponse } from '../axios/AxiosHttpClientAdapter'
 import { makePaymentBaseUrl } from './make-payment-base-url'
-import { formatDateMonthAndYear, getExpiredDate } from '../../helpers/convert-date'
+import { formatDateMonthAndYear } from '@/helpers/convert-date'
+import { getExpiredDate } from '@/helpers/payment-method'
 
 export const listPaymentService = async () => {
   let httpResponse = await AxiosHttpClientAdapter.request({
