+++ src/tests/helpers/convert-date.test.js
@@ -2,8 +2,7 @@ import { describe, it, expect } from 'vitest'
 import {
   convertValueToDate,
   convertDateToLocalTimezone,
-  formatDateMonthAndYear,
-  getExpiredDate
+  formatDateMonthAndYear
 } from '@/helpers/convert-date'
 
 describe('convertDate', () => {
@@ -40,36 +39,5 @@ describe('convertDate', () => {
     const actualDateString = formatDateMonthAndYear(month, year)
 
     expect(actualDateString).toEqual(expectedDateString)
-  })
-
-  it('Check expiration status for a future date', () => {
-    const month = 10
-    const year = 2024
-
-    const expectedStatus = ''
-    const actualStatus = getExpiredDate(month, year)
-
-    expect(actualStatus).toEqual(expectedStatus)
-  })
-
-  it('Check expiration status for a past date', () => {
-    const month = 5
-    const year = 2021
-
-    const expectedStatus = 'Expired'
-    const actualStatus = getExpiredDate(month, year)
-
-    expect(actualStatus).toEqual(expectedStatus)
-  })
-
-  it('Check expiration status for the current month and year', () => {
-    const currentDate = new Date()
-    const month = currentDate.getMonth() + 1
-    const year = currentDate.getFullYear()
-
-    const expectedStatus = ''
-    const actualStatus = getExpiredDate(month, year)
-
-    expect(actualStatus).toEqual(expectedStatus)
   })
 })
